stops guarded arm movement
